requirements.txt
httpx[http2] >= 0.16.1
PyYAML >= 5.4.1
requests >= 2.24.0
